+++ source/js/vendors.js
@@ -1,5 +1,3 @@
 //= include vendor/skip-link-focus-fix.js
-//= include vendor/masonry/imagesloaded.js
-//= include vendor/masonry/masonry.pkgd.js
 //= include vendor/bootstrap/modal.js
 //= include vendor/bootstrap/collapse.js