+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] - 2016-11-04
+### Fixed
+- General small bugs fixes.
+
+### Changed
+- Better transition effect between Load more button and loader.
+- Load more button replaced with message when there are no more posts to show.
+
+### Added
+- TinyMCE editor style
+
 ## [1.0.1] - 2016-06-13
 ### Fixed
 - Footer recent posts bug